+++ src/routes/jobs/+page.svelte
@@ -2,7 +2,7 @@
   import { fade } from 'svelte/transition';
   import { locale, t } from 'svelte-intl-precompile';
 
-  import { browser } from '$app/env';
+  import { browser } from '$app/environment';
   import { goto } from '$app/navigation';
   import { page } from '$app/stores';
